requirements.txt
aiogram==3.*
python-dotenv
gspread
google-auth
google-api-python-client
